+++ .preserve_layers/openexr/python/numpy/_typing/__init__.py
@@ -1,3 +1,0 @@
-"""Private counterpart of ``numpy.typing``."""
-
-from __future__ import annotations